Refactor Crashlytics integration
- Remove app config for Crashlytics
- Remove Firebasable module
- Increase query spans to be 30 days
- Add platform to the queries
